Remove comment about Game Over not working
Game over index was wrong in track listing, but the real reason it
wasn't playing was an error with my build scripts

Turns out in Batch if you put a REM comment at the end of a line, the
whole thing gets commented out? At least that's what was happening
here.